Download references that match a given WOS query

This function sends a query to the Web Of Science Lite API (https://developer.clarivate.com/apis/woslite) and returns references that match this query.

To learn how to write a WOS query, users can read the WOS documentation available at: https://images.webofknowledge.com/images/help/WOK/contents.html. A list of WOS field tags is available at https://images.webofknowledge.com/images/help/WOS/hs_wos_fieldtags.html.

It's strongly recommended to use the function wos_search before to have an idea on how many records you will download.

Important: Due to WOS LITE API limitations, the total number of records cannot exceed 100,000.

wos_get_records(query, database = "WOK", sleep = 1)

Arguments

query

a character of length 1. The query to send to the WOS lite API. Visit the WOS documentation at https://images.webofknowledge.com/images/help/WOK/contents.html on how to write a query.

database

a character of length 1. One among BCI (BIOSIS Citation Index), BIOABS (Biological Abstracts), BIOSIS (BIOSIS Previews), CCC (Current Contents Connect), DCI (Data Citation Index), DIIDW (Derwent Innovations Index), MEDLINE (Medline), WOS (Web of Science Core Collection), ZOOREC (Zoological Record), and WOK (all databases). Default is WOK (all databases).

sleep

an numeric of length 1. To not stress the WOS LITE API, a random number between 0 and sleep will be picked to suspend queries.

Value

A data.frame with n rows (where n is the total number of references) and the following 21 variables: ut: the unique identifier of the reference in the Web of Science system; doc_type: the document type; title: the title of the reference; authors: the authors of the reference; book_authors: the book authors (if applicable); book_group_authors: the book group authors (if applicable); keywords: the authors keywords; source: the title of the source (journal, book, etc.) in which the reference was published; volume: the volume number; issue: the issue number; pages: the pages range in the source; no_article: the article number; published_date: the published date; published_year: the published year; supplement_number: the supplement number (if applicable); special_issue: SI in case of a special issue (if applicable); book_series_title: the book series title (if applicable); doi: the Digital Object Identifier; eissn: the Electronic International Standard Identifier Number; issn: the International Standard Identifier Number; isbn: International Standard Book Number.

Examples

if (FALSE) {
## Download references of an author ----
query <- "AU=(\"Casajus N\")"
wos_search(query)
refs <- wos_get_records(query)
}