Retrieve the number of records that match a given WOS query

This function sends a query to the Web Of Science Lite API (https://developer.clarivate.com/apis/woslite) and returns the total number of records that match this query.

To learn how to write a WOS query, users can read the WOS documentation available at: https://images.webofknowledge.com/images/help/WOK/contents.html. A list of WOS field tags is available at: https://images.webofknowledge.com/images/help/WOS/hs_wos_fieldtags.html.

wos_search(query, database = "WOS")

Arguments

query

a character of length 1. The query to send to the WOS lite API. Visit the WOS documentation at: https://images.webofknowledge.com/images/help/WOK/contents.html on how to write a WOS query.

database

a character of length 1. One among BCI (BIOSIS Citation Index), BIOABS (Biological Abstracts), BIOSIS (BIOSIS Previews), CCC (Current Contents Connect), DCI (Data Citation Index), DIIDW (Derwent Innovations Index), MEDLINE (Medline), WOS (Web of Science Core Collection), ZOOREC (Zoological Record), and WOK (all databases). Default is WOS (all databases).

Value

The total number of records (integer of length 1) that match the query.

Examples

if (FALSE) {
## Search in TOPIC an exact expression ----
query <- "TS=\"salmo salar\""
wos_search(query)

## Search in TOPIC an exact expression and another term ----
query <- "TS=(\"salmo salar\" AND conservation)"
wos_search(query)

## Search for a specific year ----
query <- "TS=(\"salmo salar\" AND conservation) AND PY=2021"
wos_search(query)

## Search for a time span ----
query <- "TS=(\"salmo salar\" AND conservation) AND PY=2010-2021"
wos_search(query)

## Search for an author ----
query <- "AU=(\"Casajus N\")"
wos_search(query)
}